IGNITE-8740: Support reuse of already initialized Ignite in IgniteSpringBean
