+++ modules/spring/src/test/java/org/apache/ignite/testsuites/IgniteSpringTestSuite.java
@@ -19,6 +19,7 @@ package org.apache.ignite.testsuites;
 
 import junit.framework.TestSuite;
 import org.apache.ignite.cache.spring.GridSpringCacheManagerSelfTest;
+import org.apache.ignite.cache.spring.GridSpringCacheManagerSpringBeanSelfTest;
 import org.apache.ignite.cache.spring.SpringCacheManagerContextInjectionTest;
 import org.apache.ignite.cache.spring.SpringCacheTest;
 import org.apache.ignite.spring.injection.IgniteSpringBeanSpringResourceInjectionTest;
@@ -38,6 +39,7 @@ import org.apache.ignite.spring.IgniteStartFromStreamConfigurationTest;
 import org.apache.ignite.spring.injection.GridServiceInjectionSpringResourceTest;
 import org.apache.ignite.testframework.IgniteTestSuite;
 import org.apache.ignite.transactions.spring.GridSpringTransactionManagerSelfTest;
+import org.apache.ignite.transactions.spring.GridSpringTransactionManagerSpringBeanSelfTest;
 import org.apache.ignite.transactions.spring.SpringTransactionManagerContextInjectionTest;
 
 /**
@@ -63,6 +65,7 @@ public class IgniteSpringTestSuite extends TestSuite {
         suite.addTestSuite(GridP2PUserVersionChangeSelfTest.class);
 
         suite.addTestSuite(GridSpringCacheManagerSelfTest.class);
+        suite.addTestSuite(GridSpringCacheManagerSpringBeanSelfTest.class);
 
         suite.addTestSuite(IgniteDynamicCacheConfigTest.class);
 
@@ -76,6 +79,7 @@ public class IgniteSpringTestSuite extends TestSuite {
         suite.addTestSuite(CachePojoStoreXmlWithSqlEscapeSelfTest.class);
 
         suite.addTestSuite(GridSpringTransactionManagerSelfTest.class);
+        suite.addTestSuite(GridSpringTransactionManagerSpringBeanSelfTest.class);
 
         suite.addTestSuite(GridServiceInjectionSpringResourceTest.class);
         suite.addTestSuite(IgniteSpringBeanSpringResourceInjectionTest.class);
